Bugfixy i dodanie wyboru cwiczen z grupy
Naprawiono dodawanie grup do ćwiczeń (pola stringów), dodawanie ćwiczeń do grup (lista stringów)
Dodano możliwość wybierania ćwiczeń z wielu grup jako grupa tymczasowa.

+++ lib/Controllers/WorkoutController.dart
@@ -15,7 +15,7 @@ class WorkoutController{
 
   WorkoutController(this.selectedWorkout);
 
-  void addExerciseToWorkout(Exercise e, int series, List<int> weights,
+  void addExerciseToWorkout(Exercise e, int series, List<double> weights,
       List<int> repetitions){
     selectedWorkout.exercises.add(ExerciseWrapper.full(exercise: e, series: series,
     weights: weights, repetitions: repetitions));
